Initial push of Database branch from GitLab
This is the push as is from the GitLab repository. Remember to change the port and password information in the sql.cs file per your personal installation.

The schema (without foreign key assignments) as of last meeting is included.  In addition, there's a data dump which includes the dummy data used in testing.

TODO: Generate actual dummy data to populate the database.