Efficiency patch and introduction of Keyrword association
Both changePassword and changeUsername used an unnecessary query which opened an unnecessary connection. I consolidated the extra query into the first, and those blocks now make one less connection in the process of verifying information.

Additionally, the include function was introduced. It takes a keyword and a note_id. If the keyword doesn't exist, it will add it to the keyword table. In either case, it will attempt to build an association between the note_id provided and the appropriate entry in the keyword database as an entry in the contains database.

NOTE: This hasn't been tested.