Added includeMultiple.
Created a function for processing multiple keywords to make the work less strenuous for whoever needs to write the parsing algorithm for the notes. It may not be necessary depending on how they go about it.

Provide it a list of strings that are keywords to be associated with the note, and provide the note id. It will handle the rest.