Update NoteParser.cs
NOTE: This code has not been tested. Use at your own risk!

I went ahead and patched in some temporary code from stack overflow for scrapping all of the RTF formatting data and generating a txt file instead. Then, it parses the temporary txt file as previously defined.

There is a path to improving this that will be undertaken as time permits.