added addCourse()
addCourse creates a new course and returns the id of the newly created course.

If the course already exists, then it will return the id of the already existing course.

If any field contains an attempted SQL injection, it will fail with -3.

If it fails to create the course for any other reason, it will fail with -1.