Updated removeUser()
Added the cascade behavior to note_ids to avoid user deletion's impact on note records.

This is necessary because we generate new ids based on the number of existing users, notes, and courses.

When a user is deleted, we need to roll back all greater user_ids. Additionally, we need to modify all of the ids in the is_taking relationship to preserve the user's access to their notes. This behavior was already present.

However, when a user is deleted, their notes should also be deleted. When notes are deleted, a cascade change needs to happen both in the note table and in the is_taking table. This behavior was added.

While I was at it, I condensed the serious of actions into a single query to reduce the number of network interactions.